+++ lib/evostream/event/info.rb
@@ -6,7 +6,7 @@
 # Define constant to gem.
 module Evostream
   # Define version to gem
-  VERSION = '0.2.7'
+  VERSION = '0.3.0'
 
   # Name to gem
   GEM_NAME = 'evostream-event'