+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [0.2.7] - 2017-05-16
 ### [Fixed]
+- Change method encode base 64, use strict instead of urlsafe
 - Use Hash response with symbol
 - Refractoring response
 - ~~Fix json response~~