+++ CHANGELOG.md
@@ -5,9 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [0.2.7] - 2017-05-16
+## [0.3.0] - 2017-05-16
 ### [Added]
 - Add command listConfig
+- Add command listStreams
 
 ### [Fixed]
 - Change method encode base 64, use strict instead of urlsafe